Organized, TODO readme update
