+++ .gitignore
@@ -15,6 +15,7 @@ VTS/Core/**/*.meta
 LICENSE.meta
 VTS.meta
 Examples.meta
+Examples/**/*.meta
 README.md.meta
 # Asset meta data should only be ignored when the corresponding asset is also ignored
 !/[Aa]ssets/**/*.meta
