--- Examples/Unity.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: b491947e9ac5f9442a2ec0518449670a
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
